refactor: 책 검색 이벤트리스너 분리
